Day 7. This time no special insight was needed, a brute force solution resolved both part "a" and part "b" quickly.
I think that was the point: in yesterday's problem optimization was needed to solve part "b," but in today's problem no optimization was necessary (and perhaps was not possible, or I'm not mathematically inclined enough).

Donald Knuth said that "premature optimization is the root of all evil," partly because it can introduce complexity, but also because it can lock you down to a solution that isn't flexible to even slight variations on the problem. I think that may have been the message here.

But it's always fun to play with the problem anyway: am I missing something? Is this simpler than it looks?